Refactor models into one file per model and delete the old models monolith
