+++ natlas-server/app/models/__init__.py
@@ -1,0 +1,23 @@
+from app.models.agent import Agent
+from app.models.agent_config import AgentConfig
+from app.models.agent_script import AgentScript
+from app.models.config_item import ConfigItem
+from app.models.email_token import EmailToken
+from app.models.natlas_services import NatlasServices
+from app.models.rescan_task import RescanTask
+from app.models.scope_item import ScopeItem
+from app.models.tag import Tag
+from app.models.user import User
+
+__all__ = [
+	'Agent',
+	'AgentConfig',
+	'AgentScript',
+	'ConfigItem',
+	'EmailToken',
+	'NatlasServices',
+	'RescanTask',
+	'ScopeItem',
+	'Tag',
+	'User'
+]
